Move redis too
